fix(ci): python-coverage-shards — split test_pr_monitor_operator_hints.py under 1500-line limit
The maintainability guardrail also flagged tests/unit/runtime/test_pr_monitor_operator_hints.py
at 2870 lines. Move the grant/resume slice (the _seed_active_grant / _set_block_resume_phase
helpers and their tests) into a new test_pr_monitor_operator_hints_part_002.py sibling, sharing
the same imports/fixtures preamble; both files now sit under the limit. Also re-point the
_PROTECTED_BLOCK_PRESERVED_HEAD_STATE_KEY import at its canonical home (awf.runtime.pr_monitor),
since the symbol it was re-imported through (remote_repair) no longer re-exports it.
